refactor(imports): Update import paths after folder structure flattening
- Adjusted import statements in ViewModel and provider files
- Reflects relocation of models, repositories, and validators
- Ensures consistency across new directory structure
