+++ lib/ui/weather/view_model/async_weather_view_model.dart
@@ -1,8 +1,8 @@
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 
 import '../../../../core/logger/app_log.dart';
-import '../../../../data/repositories/weather/repositories/unified_weather_repository_provider.dart';
-import '../../../../domain/models/weather/models/forecast.dart';
+import '../../../data/repositories/weather/unified_weather_repository_provider.dart';
+import '../../../domain/models/weather/forecast.dart';
 
 class AsyncWeatherViewModel extends AsyncNotifier<Forecast> {
   @override
