+++ lib/ui/weather/view_model/providers/async_weather_view_model_provider.dart
@@ -1,6 +1,6 @@
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 
-import '../../../../../domain/models/weather/models/forecast.dart';
+import '../../../../domain/models/weather/forecast.dart';
 import '../async_weather_view_model.dart';
 
 final asyncWeatherViewModelProvider =
